+++ src/fevor_distribution.hh
@@ -48,7 +48,7 @@ class fevor_distribution {
         
     // functions
         // preform a time step
-        std::vector<double> stepInTime(const double &temperature, const std::vector<double> &stress, double &modelTime, const double &timeStep, double &nMigre, double &nPoly, std::vector<double> &bulkEdot);
+        std::vector<double> stepInTime(const double &temperature, const std::vector<double> &stress, double &modelTime, const double &timeStep, unsigned int &nMigre, unsigned int &nPoly, std::vector<double> &bulkEdot);
         
         // calculate NNI softness parameter
         void getSoftness(std::vector<std::vector<double>> &crystalM, std::vector<double> &bulkM, std::vector<double> &bulkEdot, const std::vector<double> &stress);
