+++ test/test_distribution.cc
@@ -50,7 +50,7 @@ void test_stepInTime() {
                                    10000,     0,-10000};
     double modelTime = 0.0;
     double timeStep = 1000.0*365.0*24.0*60.0*60.0;
-    double nMigre, nPoly;
+    unsigned int nMigre, nPoly;
     nMigre = nPoly = 0;
     std::vector<double> bulkEdot= { 0.0, 0.0, 0.0,
                                     0.0, 0.0, 0.0,
